% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignDomains.R
\name{alignDomainsRigid}
\alias{alignDomainsRigid}
\title{Independently align multiple structure domains rigidly.}
\usage{
alignDomainsRigid(fixed, mobile, domains)
}
\arguments{
\item{fixed}{A protein structure of class "pdb" (from bio3d) to
act as the fixed template.}

\item{mobile}{A protein structure of class "pdb" (from bio3d) which is
aligned to the fixed structure.}

\item{domains}{The locations of domains to align. Provided as a list of
lists, where each outer item describes one domain and each inner item is
a two-item list containing an integer vector of the domain's indices in
the fixed protein and an integer vector of the domain's indices in the
mobile protein. Each pair must contain the same number of residues across
structures.}
}
\value{
A protein structure of class "pdb" with the residues from the mobile
structure rigidly aligned to the fixed structure on each domain.
}
\description{
A function to align a mobile protein structure against a fixed template
protein structure by independently aligning each pair of provided domains.
The alignment is rigid, eg the domain is moved as one unit. Further,
the locations of atoms in the mobile protein structure which are not part of
a domain do not move.
}
\examples{
# library(bio3d)
# Open the predicted structure 6ofs_predicted.pdb provided by the package.
# This is the predicted structure of an E coli zinc protease.
predicted6ofsFile <- system.file("extdata", "6ofs_predicted.pdb",
                                 package = "PDBCleanup")
predicted6ofs <- bio3d::read.pdb(predicted6ofsFile)
# Open the experimental structure 6ofs_experimental.pdb provided by the
# package. This is the experimentally determined structure of the same E coli
# zinc protease.
experimental6ofsFile <- system.file("extdata", "6ofs_experimental.pdb",
                                 package = "PDBCleanup")
experimental6ofs <- bio3d::read.pdb(experimental6ofsFile)

# Define the domains. Here, they are the same because we use alternate
# structures from one species rather than structures from different species
domains6ofs <- list(list(36:486, 36:486), list(511:930, 511:930))

# The domains are placed differently between structures. This alignment moves
# the predicted domains into the places predicted by the experimental
# domains, which is beneficial if the experimental domain locations is better.
alignedPredicted6ofs <- alignDomainsRigid(experimental6ofs, predicted6ofs,
                                          domains6ofs)

}
\references{
Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.
}
