% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProteinQuality.R
\name{plotProteinQuality}
\alias{plotProteinQuality}
\title{Plot the quality of a protein structure along its sequence.}
\usage{
plotProteinQuality(structure, title, predicted)
}
\arguments{
\item{structure}{A protein structure of class "pdb" (from bio3d) to
visualize B-values from.}

\item{title}{The title for the generated plot.}

\item{predicted}{Whether the protein structure is predicted, so the y axis
label is "Predicted Structure Quality," or experimental, so the y axis
label is "Experimental B-factor."}
}
\value{
The histogram plot of the B-factor along the protein sequence.
}
\description{
A function to visualize the B-factor of a protein structure as a bar plot
along the sequence. The user specifies whether the structure is predicted
to specify the y-axis label and gives a custom plot title.
}
\examples{
# library(bio3d)
# library(ggplot2)
# Load and visualize the quality along an AlphaFold prediction.
predicted6ofsFile <- system.file("extdata", "6ofs_predicted.pdb",
                                 package = "PDBCleanup")
predicted6ofs <- bio3d::read.pdb(predicted6ofsFile)
plotProteinQuality(predicted6ofs,
                   title = "Quality of Predicted 6ofs structure",
                   predicted = TRUE)
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag
New York, 2016.
}
