% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Launch Shiny App for PDBCleanup}
\usage{
runShiny()
}
\value{
No return value. Instead, open a Shiny page.
}
\description{
A function that launches the Shiny app for PDBCleanup.
This app provides a web interface for the package functions
selectHighResolution, selectOrdered, and alignDomains from a user-uploaded
file. The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
PDBCleanup:runShiny()
}

}
\references{
Chang W, Cheng J, Allaire J, Sievert C, Schloerke B, Xie Y, Allen J,
McPherson J, Dipert A, Borges B (2023).
shiny: Web Application Framework for R.
R package version 1.8.0, https://CRAN.R-project.org/package=shiny>

Based on:
Silva, A. (2022) TestingPackage: An Example R Package For BCB410H.
Unpublished. URL https://github.com/anjalisilva/TestingPackage.
}
