% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectHighResolution.R
\name{selectHighResolution}
\alias{selectHighResolution}
\title{Select High Resolution Regions of a Protein Structure}
\usage{
selectHighResolution(structure, predicted, threshold = NA)
}
\arguments{
\item{structure}{A protein structure of class "pdb" (bio3d format) to extract
high resolution regions from.}

\item{predicted}{A boolean indicating whether the provided structure is a
predicted structure. If TRUE, select regions with high B-factor as
high resolution. If FALSE, select regions with low B-factor as high
resolution.
Defaults to TRUE.}

\item{threshold}{An optional parameter specifying the threshold
value to identify high resolution regions. If predicted is TRUE, defaults
to 70 per AlphaFold guidelines. If predicted is FALSE, defaults to
60, which corresponds to 1.5 A mean deviation.}
}
\value{
A protein structure of class "pdb" containing only high resolution
regions which satisfy the specified B-factor requirement.
}
\description{
A function to select high resolution regions in a protein structure.
High resolution regions are selected using B-factor.
The user provides whether the structure is an AlphaFold predicted structure,
meaning high B-factor indicates high resolution, or an experimental
structure, meaning low B-factor indicates high resolution. They can
optionally provide a threshold value, otherwise the function uses a default
threshold.
}
\examples{
# library(bio3d)
# Select high quality regions in an AlphaFold prediction.
predicted6ofsFile <- system.file("extdata", "6ofs_predicted.pdb",
                                 package = "PDBCleanup")
predicted6ofs <- bio3d::read.pdb(predicted6ofsFile)
# Use default threshold of >= 70 for predicted structures
predicted6ofsHighRes <- selectHighResolution(structure = predicted6ofs,
                                             predicted = TRUE)

# Select high quality regions in an experimental prediction.
experimental6ofsFile <- system.file("extdata", "6ofs_experimental.pdb",
                                    package = "PDBCleanup")
experimental6ofs <- bio3d::read.pdb(experimental6ofsFile)
# Use custom threshold of <= 80, or < 1.75 A mean deviation
experimental6ofsHighRes <- selectHighResolution(structure = experimental6ofs,
                                                predicted = FALSE,
                                                threshold = 80)

}
\references{
Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Jumper, J., Evans, R., Pritzel, A. et al.
Highly accurate protein structure prediction with AlphaFold.
Nature 596, 583–589 (2021).
https://doi.org/10.1038/s41586-021-03819-2

Zhoutong Sun, Qian Liu, Ge Qu, Yan Feng, and Manfred T. Reetz.
Utility of B-Factors in Protein Science: Interpreting Rigidity,
Flexibility, and Internal Motion and Engineering Thermostability.
Chemical Reviews 2019 119 (3), 1626-1665.
https://doi.org/10.1021/acs.chemrev.8b00290
}
