% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectStructure.R
\name{selectOrdered}
\alias{selectOrdered}
\title{Select ordered regions of a protein structure.}
\usage{
selectOrdered(structure, threshold = 0, windowSize = 21)
}
\arguments{
\item{structure}{A bio3d pdb protein structure which the function will
select the ordered residues from.}

\item{threshold}{An optional numerical value specifying the threshold for
ordered regions. Defaults to scores greater than 0 denoting order.}

\item{windowSize}{An optional odd integer value specifying the window size
used when calculating disorder score. Defaults to 21.}
}
\value{
A bio3d pdb protein structure containing only ordered residues with
disorder scores satisfying the specified threshold.
}
\description{
A function to select all intrinsically ordered residues of a protein
structure and remove all intrinsically disordered residues. Residues are
classified using the FoldIndex score (implemented in the idpr package rather
than re-implementing), which predicts disorder at each residue with the
formula D = 2.785 H - | R | - 1.151 where H is the scaled mean hydropathy
and R is the scaled mean charge within a sequence window around the residue.
It ranges from -1 to 1 where positive scores indicate ordered residues while
negative scores indicate disordered residues.
The user optionally provides a threshold (default 0) to denote order and
optionally provides a windowSize (default 21) for this calculation.
}
\details{
The first and last windowSize // 2 residues are removed, as they lack a
sufficient window size to calculate FoldIndex score.
}
