% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectStructure.R
\name{selectResno}
\alias{selectResno}
\title{Select the residues of a protein structure at the provided indices.}
\usage{
selectResno(structure, resno)
}
\arguments{
\item{structure}{A bio3d pdb protein structure which the function will
select the specified residues from.}

\item{resno}{A vector of integers which are the residues to select.}
}
\value{
A bio3d pdb protein structure containing only the specified residues.
}
\description{
A function to select the specified residues of a given protein structure and
return a modified structure with all other residues removed.
This is primarily a helper function for the functions selectHighResPredicted
and selectHighResExperimental. However, it is exported if users have more
general applications.
}
\examples{
# library(bio3d)
# Open a PDB structure. The package provides one at 6ofs_experimental.pdb,
# which is the experimentally determined structure of an E coli zinc protease.
experimental6ofsFile <- system.file("extdata", "6ofs_experimental.pdb",
                                    package = "PDBCleanup")
experimental6ofs <- bio3d::read.pdb(experimental6ofsFile)
# Define the residues of interest. These are the two structured domains.
domainsResno6ofs <- c(36:486, 511:930)
# Select the domains from the full 6ofs structure
domains6ofs <- selectResno(structure = experimental6ofs,
                           resno = domainsResno6ofs)

}
\references{
Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.
}
