% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewStructure.R
\name{viewStructure}
\alias{viewStructure}
\title{Visualize the 3D structure of a protein.}
\usage{
viewStructure(structure)
}
\arguments{
\item{structure}{A protein structure of class "pdb" (from Bio3d) to
visualize as a 3D model.}
}
\value{
The interactive viewer showing the 3D structure model. Shown in the
viewer in RStudio, otherwise shown in a browser window.
}
\description{
A function to visualize the provided structure with a 3D cartoon model.
This function opens a browser window with the interactive 3D model.
}
\examples{
# library(bio3d)
# Load and visualize an AlphaFold prediction.
6ofsPredictedFile <- system.file("extdata", "6ofs_predicted.pdb",
                                 package = "PDBCleanup")
6ofsPredicted <- bio3d::read.pdb(6ofsPredictedFile)
viewStructure(6ofsPredicted)

}
\references{
Su W, Johnston B (2021). \emph{r3dmol: Create Interactive 3D Visualizations of
Molecular Data}. R package version 0.1.2,
\url{https://CRAN.R-project.org/package=r3dmol}.
}
