% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProteinQuality.R
\name{plotProteinQuality}
\alias{plotProteinQuality}
\title{Plot the quality of a protein structure along its sequence.}
\usage{
plotProteinQuality(
  structure,
  title,
  xtitle = "Sequence Position",
  ytitle = "Predicted Sequence Quality"
)
}
\arguments{
\item{structure}{A bio3d pdb protein structure to plot quality values from.}

\item{title}{A character string containing the title for the generated plot.}

\item{xtitle}{An optional character string containing the x-axis title.
Default value is "Sequence Position"}

\item{ytitle}{An optional character string containing the y-axis title.
Default value is "Predicted Structure Quality"}
}
\value{
The histogram plot of the B-factor along the protein sequence.
}
\description{
A function to visualize the data in the B-factor column of a protein
structure as a bar plot along the sequence. The user specifies the plot title
and optionally the x-axsis and y-axis title.
}
\examples{
# library(bio3d)
# library(ggplot2)
# Open the predicted structure 6ofs_predicted.pdb provided by the package.
# This is the predicted structure of an E coli zinc protease.
predicted6ofsFile <- system.file("extdata", "6ofs_predicted.pdb",
                                 package = "PDBCleanup")
predicted6ofs <- bio3d::read.pdb(predicted6ofsFile)
plotProteinQuality(predicted6ofs,
                   title = "Quality of Predicted 6ofs structure")
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag
New York, 2016.
}
