% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectHighResolution.R
\name{selectHighResolution}
\alias{selectHighResolution}
\title{Select high resolution regions of a protein structure.}
\usage{
selectHighResolution(structure, predicted, threshold = NA)
}
\arguments{
\item{structure}{A bio3d pdb protein structure which the function will
select the specified residues from.}

\item{predicted}{A logical value indicating if the structure is predicted or
experimental. Predicted structures are interpreted using B-factor (lower
B-factor is high resolution) and experimental structures are interpreted
using pLDDT (higher pLDDT is high resolution).}

\item{threshold}{An optional numeric value specifying the threshold for high
resolution regions. In predicted structures, the default is 70 and all
residues with higher pLDDT are high resolution. In experimental
structures, the default is 100 (~2.75 A uncertainty) and all residues
with lower B-factor are marked high resolution.}
}
\value{
A bio3d pdb protein structure containing only high-quality residues
satisfying the specified threshold.
}
\description{
A function to select high resolution regions in a protein structure.
}
\details{
In an experimental structure, these regions are determined using the
B-factor, which measures the average uncertainty in atom location using the
formula B = 4*pi^2 u^2 / 3 where u is the uncertainty in angstroms. A low
B-factor indicates a higher quality region, and the function selects all
residues containing atoms with a B-factor below a threshold.

In a predicted structure, these regions are determined using the pLDDT
stored in the B-factor column, which is AlphaFold's prediction for that
residue's score on the local distance difference test. A high pLDDT
indicates the predicted residue is likely close to the true position, while
a low pLDDT incidates less confidence in the predicted residue location.
The function selects all residues with pLDDT above a threshold.
}
\examples{
# library(bio3d)
# Open the predicted structure 6ofs_predicted.pdb provided by the package.
# This is the predicted structure of an E coli zinc protease.
predicted6ofsFile <- system.file("extdata", "6ofs_predicted.pdb",
                                    package = "PDBCleanup")
predicted6ofs <- bio3d::read.pdb(predicted6ofsFile)
# Use default threshold of pLDDT >= 70
predicted6ofsHighRes <- selectHighResolution(structure = predicted6ofs,
                                             predicted = TRUE)

# Open the experimental structure 6ofs_experimental.pdb provided by the
# package. This is the experimentally determined structure of the same E coli
# zinc protease.
experimental6ofsFile <- system.file("extdata", "6ofs_experimental.pdb",
                                    package = "PDBCleanup")
experimental6ofs <- bio3d::read.pdb(experimental6ofsFile)
# Use custom threshold of bfactor <= 80, or under ~2.5 A mean deviation
experimental6ofsHighRes <- selectHighResolution(structure = experimental6ofs,
                                                predicted = FALSE,
                                                threshold = 80)

}
\references{
Grant, B.J. et al. (2006) Bioinformatics 22, 2695--2696.

Zhoutong Sun, Qian Liu, Ge Qu, Yan Feng, and Manfred T. Reetz.
Utility of B-Factors in Protein Science: Interpreting Rigidity,
Flexibility, and Internal Motion and Engineering Thermostability.
Chemical Reviews 2019 119 (3), 1626-1665.
https://doi.org/10.1021/acs.chemrev.8b00290
}
