% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewStructure.R
\name{viewStructure}
\alias{viewStructure}
\title{Visualize the 3D structure of a protein.}
\usage{
viewStructure(structure, color = "#00cccc")
}
\arguments{
\item{structure}{A bio3d pdb protein structure to visualize as a 3D model.}

\item{color}{An optional character string specifying the color for the 3D
model. Default value is "#00cccc".}
}
\value{
The interactive viewer showing the 3D structure model. Shown in the
viewer in RStudio, otherwise shown in a browser window.
}
\description{
A function to visualize the provided structure with a 3D cartoon model.
This function opens a browser window with the interactive 3D model.
}
\examples{
# library(bio3d)
# Load and visualize an AlphaFold prediction.
predicted6ofsFile <- system.file("extdata", "6ofs_predicted.pdb",
                                 package = "PDBCleanup")
predicted6ofs <- bio3d::read.pdb(predicted6ofsFile)
viewStructure(predicted6ofs)

}
\references{
Su W, Johnston B (2021). r3dmol: Create Interactive 3D Visualizations of
Molecular Data. R package version 0.1.2,
\url{https://CRAN.R-project.org/package=r3dmol}.
}
