% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewStructure.R
\name{viewStructure}
\alias{viewStructure}
\title{Visualize the 3D structure of a protein.}
\usage{
viewStructure(structure)
}
\arguments{
\item{structure}{A protein structure of class "pdb" (from Bio3d) to
visualize as a 3D model.}
}
\value{
NULL. As a side effect, a browser window is opened with an
interactive interface showing the 3D structure model.
}
\description{
A function to visualize the provided structure with a 3D cartoon model.
This function opens a browser window with the interactive 3D model.
}
\references{
Su W, Johnston B (2021). \emph{r3dmol: Create Interactive 3D Visualizations of
Molecular Data}. R package version 0.1.2,
\url{https://CRAN.R-project.org/package=r3dmol}.
}
